% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_na}
\alias{all_na}
\title{Check if each element of a vector is NA}
\usage{
all_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector
}
\description{
Check if each element of a vector is NA
}
\examples{
all_na(c(1, NA, 3))
all_na(c(NA, NA, NA, NA))
}
