% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_bbd.R
\name{anova_bbd}
\alias{anova_bbd}
\title{ANOVA do modelo Box-Behnken}
\usage{
anova_bbd(fit)
}
\arguments{
\item{fit}{Objeto ajustado por \code{bbd_fit()}.}
}
\value{
data.frame com a tabela ANOVA.
}
\description{
Retorna a tabela de análise de variância do modelo ajustado.
}
