% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-reference.R
\name{bref_war_daily_batter}
\alias{bref_war_daily_batter}
\title{Download daily updated WAR data for batters from Baseball-Reference}
\usage{
bref_war_daily_batter()
}
\value{
An object with class \code{c("tbl_df", "tbl", "data.frame")} containing
batter WAR data.
}
\description{
Download daily updated WAR data for batters from Baseball-Reference
}
\examples{
\dontrun{
bref_war_daily_batter()
}
}
