% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_nonempty_df}
\alias{check_nonempty_df}
\title{Check for non-empty data frame}
\usage{
check_nonempty_df(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical vector
}
\description{
Check for non-empty data frame
}
\examples{
check_nonempty_df(data.frame())
check_nonempty_df(data.frame(1))
}
