% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeficientes_bbd.R
\name{coeficientes_bbd}
\alias{coeficientes_bbd}
\title{Coeficientes do modelo Box-Behnken}
\usage{
coeficientes_bbd(fit)
}
\arguments{
\item{fit}{Objeto ajustado por \code{bbd_fit()}.}
}
\value{
data.frame com estimativas, erro padrão, estatística t e p-valor.
}
\description{
Retorna a tabela de coeficientes do modelo ajustado.
}
