% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coltypes}
\alias{coltypes}
\title{Get the type of each column of a data frame}
\usage{
coltypes(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A character vector
}
\description{
Get the type of each column of a data frame
}
