% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contorno_dbb.R
\name{contorno_dbb}
\alias{contorno_dbb}
\title{Grafico de contorno para Box Behnken}
\usage{
contorno_dbb(fit, x, y, z.fix = 0, nlevels = 10)
}
\arguments{
\item{fit}{objeto retornado por dbb_fit()}

\item{x}{nome do fator para eixo x}

\item{y}{nome do fator para eixo y}

\item{z.fix}{valor fixo do terceiro fator, default = 0}

\item{nlevels}{numero de curvas de nivel}
}
\description{
Grafico de contorno para Box Behnken
}
