% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{correlations}
\alias{correlations}
\title{Test of correlation for multiple variables between paired samples}
\usage{
correlations(data, x.variables, y.variables = NULL, parallel = TRUE,
  parallel.inorder = FALSE, ...)
}
\arguments{
\item{data}{a data frame with the data for all variables to be correlated.}

\item{x.variables}{character vector of column names for variables to correlate to \code{variables.y}. If \code{variables.y} is left empty, correlate all \code{variables.x} to each other.}

\item{y.variables}{character vector of column names for variables to correlate to \code{variables.x}. If left \code{NULL}, correlate all \code{variables.x} to each other.}

\item{parallel}{logical flag indicating if the tasks should be run in parallel. The default value is \code{TRUE}.}

\item{parallel.inorder}{logical flag indicating if parallelry run tasks should be combined in the same order that they were submitted. If the order is not important, then setting \code{parallel.inorder} to \code{FALSE} can result in improved performance. The default value is \code{FALSE}.}

\item{...}{further arguments to be passed to \code{cor.test} method (e.g. \code{method}, \code{conf.level}, \code{alternative}, \code{exact}).}
}
\value{
A data.frame with correlation results for each pair.
}
\description{
Test of correlations for multiple variables between paired samples, using \code{cor.test} function.
}
\examples{
correlations(data=iris, x.variables=c("Sepal.Length", "Sepal.Width",
             "Petal.Length", "Petal.Width"), parallel=FALSE)

correlations(data=iris, x.variables=c("Sepal.Length", "Sepal.Width"),
             y.variables=c("Petal.Length", "Petal.Width"), parallel=FALSE)

}
\seealso{
\code{\link{cor.test}} for correlation testing function and its arguments.
}

