% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar_relatorio_bbd.R
\name{exportar_relatorio_bbd}
\alias{exportar_relatorio_bbd}
\title{Exporta relatório completo do BBD}
\usage{
exportar_relatorio_bbd(
  fit,
  arquivo = "Relatorio_BBD",
  formato = "txt",
  tipo_otimo = NULL
)
}
\arguments{
\item{fit}{objeto da classe bbd_fit}

\item{arquivo}{nome do arquivo a ser salvo, sem extensão}

\item{formato}{"txt" ou "doc"}

\item{tipo_otimo}{NULL para não incluir ótimo, "min" para minimizar
a resposta ou "max" para maximizar}
}
\value{
invisivelmente, o caminho completo do arquivo gerado
}
\description{
Exporta relatório completo do BBD
}
