% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{get_statcast}
\alias{get_statcast}
\title{Download all Statcast data within a range of dates}
\usage{
get_statcast(start = Sys.Date() - 8, end = Sys.Date() - 1, ...)
}
\arguments{
\item{start}{Start date of search}

\item{end}{End data of search}

\item{...}{Additional arguments passed to \code{get_statcast_day}}
}
\value{
A \code{data.table}
}
\description{
Download all Statcast data within a range of dates
}
