% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{get_statcast_day}
\alias{get_statcast_day}
\title{Download all Statcast data for a given date}
\usage{
get_statcast_day(date = Sys.Date() - 1, verbose = FALSE)
}
\arguments{
\item{date}{A date}

\item{verbose}{Controls messaging to user}
}
\value{
A \code{data.table}
}
\description{
Download all Statcast data for a given date
}
