% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_clipboard_bbd.R
\name{ler_clipboard_bbd}
\alias{ler_clipboard_bbd}
\title{Ler dados do clipboard}
\usage{
ler_clipboard_bbd()
}
\value{
data.frame com os dados importados
}
\description{
Importa dados copiados do Excel para o R.
}
