% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_date}
\alias{make_date}
\title{Format character input as a Date}
\usage{
make_date(x)
}
\arguments{
\item{x}{\verb{[character]} A string of the form \code{"YYYY-MM-DD"}}
}
\value{
A object with class \code{"Date"}
}
\description{
Format character input as a Date
}
