% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_tibble_class}
\alias{make_tibble_class}
\title{Obtain vector of classes to create a tibble}
\usage{
make_tibble_class()
}
\value{
A character vector
}
\description{
Obtain vector of classes to create a tibble
}
