% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{melt_summarized_data}
\alias{melt_summarized_data}
\title{Convert summarize_data object into a molten data frame.}
\usage{
melt_summarized_data(x, parallel = FALSE)
}
\arguments{
\item{x}{summarized.data object to melt}

\item{parallel}{logical flag indicating if the tasks should be run in parallel. The default value is \code{FALSE}.}
}
\value{
molten \code{data.frame}
}
\description{
This function melts summarize_data objects into a long-format \code{data.frame}
for easier processing and visualization.
}
\examples{
# Add random errors to data
tmp <- mtcars
tmp[sample(nrow(tmp),sample(1:10, 1)), sample(ncol(tmp), 3)] <- NA
tmp[sample(nrow(tmp),sample(1:10, 1)), sample(ncol(tmp), 2)] <- ""

s <- summarize_data(tmp)
m <- melt_summarized_data(s)

\dontrun{

# Visualize the classes and warnings with histograms
library(ggplot2)
p <- ggplot(data=m, aes(x=Column)) + geom_histogram(col="black", aes(fill=Type))
p <- p + facet_wrap(facets = "Label", ncol=1, scales="free_y")
p <- p + theme_minimal() + scale_fill_manual(values=c("steelblue", "sienna1"))
p
}
}

