% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_bbd.R
\name{metricas_bbd}
\alias{metricas_bbd}
\title{Métricas do ajuste Box-Behnken}
\usage{
metricas_bbd(fit)
}
\arguments{
\item{fit}{Objeto ajustado por \code{bbd_fit()}.}
}
\value{
data.frame com métricas do modelo.
}
\description{
Retorna métricas principais do modelo ajustado.
}
