% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbam.R
\name{mlb_roster}
\alias{mlb_roster}
\title{Get MLB active rosters}
\usage{
mlb_roster(team)
}
\arguments{
\item{team}{\verb{[numeric]} MLB \href{https://statsapi.mlb.com/}{\code{statsapi}} team ID}
}
\value{
An object with class \code{c("tbl_df", "tbl", "data.frame")}. Specifically,
a data frame with \code{40} rows and \code{11} columns:
\describe{
\item{\code{JerseyNumber}}{\verb{[character]} desc}
\item{\code{TeamID}}{\verb{[integer]} desc}
\item{\code{PersonID}}{\verb{[integer]} desc}
\item{\code{FullName}}{\verb{[character]} desc}
\item{\code{PersonLink}}{\verb{[character]} desc}
\item{\code{PositionCode}}{\verb{[character]} desc}
\item{\code{PositionName}}{\verb{[character]} desc}
\item{\code{PositionType}}{\verb{[character]} desc}
\item{\code{PositionAbbreviation}}{\verb{[character]} desc}
\item{\code{StatusCode}}{\verb{[character]} desc}
\item{\code{StatusDescription}}{\verb{[character]} desc}
}
}
\description{
Get MLB active rosters
}
