% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nonempty_df}
\alias{nonempty_df}
\title{Check for non-empty data frame}
\usage{
nonempty_df(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical vector
}
\description{
Check for non-empty data frame
}
