% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otimo_bbd.R
\name{otimo_bbd}
\alias{otimo_bbd}
\title{Encontrar ponto otimo previsto}
\usage{
otimo_bbd(fit, objetivo = "max")
}
\arguments{
\item{fit}{objeto retornado por bbd_fit()}

\item{objetivo}{objetivo da otimizacao: "max" para maximizar ou "min" para minimizar}
}
\description{
Encontrar ponto otimo previsto
}
