% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_dbb.R
\name{pareto_dbb}
\alias{pareto_dbb}
\title{Grafico de Pareto dos efeitos DBB}
\usage{
pareto_dbb(fit, alpha = 0.05)
}
\arguments{
\item{fit}{objeto retornado por dbb_fit()}

\item{alpha}{nivel de significancia}
}
\description{
Grafico de Pareto dos efeitos DBB
}
