% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people}
\alias{people}
\title{People Information}
\format{
\subsection{\code{people}}{

A data frame with 100,715 rows and 9 columns:
\describe{
\item{name}{name Player name}
...
}
}
}
\usage{
people
}
\description{
TODO: put description here
}
\keyword{datasets}
