% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people_chadwick}
\alias{people_chadwick}
\title{Create full Chadwick people table}
\usage{
people_chadwick()
}
\value{
A table with class \code{\link[=table_class]{table_class()}} that contains all people data from
\href{https://github.com/chadwickbureau/register}{\code{chadwickbureau/register}}.
}
\description{
Create full Chadwick people table
}
