% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_sc}
\alias{people_sc}
\title{Statcast People Information}
\format{
\subsection{\code{people_sc}}{

A data frame with 100.965 rows and 2 columns:
\describe{
\item{name}{name Player name}
...
}
}
}
\usage{
people_sc
}
\description{
TODO: put description here
}
\keyword{datasets}
