% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people_search}
\alias{people_search}
\title{Fuzzy lookup within \link{people} table}
\usage{
people_search(x)
}
\arguments{
\item{x}{\verb{[character]} A character string to be searched within the names in
the \link{people} table}
}
\value{
The \link{people} table subset to rows that fuzzy match \code{x}
}
\description{
Fuzzy lookup within \link{people} table
}
\examples{
people_search("Frank Thomas")
people_search("ichiro")
}
