% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{print.summarized.data}
\alias{print.summarized.data}
\title{Printing summarize_data}
\usage{
## S3 method for class 'summarized.data'
}
\arguments{
\item{x}{object of class \code{summarize_data}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a \code{summarized.data} object.
}
\examples{
print(summarize_data(iris))

}
\seealso{
\code{\link{summarize_data}}
}

