% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clipboard_dbb.R
\name{read_clipboard_dbb}
\alias{read_clipboard_dbb}
\title{Ler dados Box Behnken copiados do Excel}
\usage{
read_clipboard_dbb()
}
\value{
data.frame
}
\description{
Ler dados Box Behnken copiados do Excel
}
