% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_bbd.R
\name{relatorio_bbd}
\alias{relatorio_bbd}
\title{Relatorio estatistico completo no padrao brasileiro}
\usage{
relatorio_bbd(fit)
}
\arguments{
\item{fit}{objeto ajustado por bbd_fit()}
}
\value{
lista com ANOVA, coeficientes, R2 e R2 ajustado formatados
}
\description{
Relatorio estatistico completo no padrao brasileiro
}
