% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_dbb.R
\name{relatorio_dbb}
\alias{relatorio_dbb}
\title{Relatorio estatistico completo no padrao brasileiro}
\usage{
relatorio_dbb(fit)
}
\arguments{
\item{fit}{objeto ajustado por dbb_fit()}
}
\value{
lista com ANOVA, coeficientes, R2 e R2 ajustado formatados
}
\description{
Relatorio estatistico completo no padrao brasileiro
}
