% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_bbd}
\alias{statcast_bbd}
\title{Download all Statcast data within a range of dates in the opinionated \code{bbd} style}
\usage{
statcast_bbd(start = Sys.Date() - 1, end = NULL, verbose = FALSE)
}
\arguments{
\item{start}{Start date of search.}

\item{end}{End data of search. Defaults to \code{NULL}, which stops search at\code{start} date.}

\item{verbose}{Controls messaging to the user. Defaults to \code{FALSE} which
provides no message. When \code{TRUE}, informs user when each \code{date} begins
downloading.}
}
\value{
An object with class \code{c("tbl_df", "tbl", "data.frame")} containing
all Statcast events between the \code{start} date and \code{end} date inclusive.
}
\description{
Download all Statcast data within a range of dates in the opinionated \code{bbd} style
}
