% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_day}
\alias{statcast_day}
\title{Download all Statcast data for a particular date}
\usage{
statcast_day(
  date = Sys.Date() - 1,
  batter = NULL,
  pitcher = NULL,
  verbose = FALSE
)
}
\arguments{
\item{date}{A date}

\item{batter}{Statcast player ID for batter of interest. Defaults to \code{NULL}.}

\item{pitcher}{Statcast player ID for pitcher of interest. Defaults to \code{NULL}.}

\item{verbose}{Controls messaging to the user. Defaults to \code{FALSE} which
provides no message. When \code{TRUE}, informs user when each \code{date} begins
downloading.}
}
\value{
A \code{data.frame} containing all Statcast events on \code{date}.
}
\description{
Download all Statcast data for a particular date
}
