% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_get_col_types}
\alias{statcast_get_col_types}
\title{Obtain vector of expected Statcast column types}
\usage{
statcast_get_col_types()
}
\value{
A character vector of the expected Statcast column types
}
\description{
Obtain vector of expected Statcast column types
}
