% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast-helpers.R
\name{statcast_get_colclasses}
\alias{statcast_get_colclasses}
\title{Obtain vector of expected Statcast column classes}
\usage{
statcast_get_colclasses()
}
\value{
A character vector of the expected Statcast column classes
}
\description{
Obtain vector of expected Statcast column classes
}
