% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast-helpers.R
\name{statcast_get_colnames}
\alias{statcast_get_colnames}
\title{Obtain vector of expected Statcast column names}
\usage{
statcast_get_colnames()
}
\value{
A character vector of the expected Statcast column names
}
\description{
Obtain vector of expected Statcast column names
}
