% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast-helpers.R
\name{statcast_get_coltypes}
\alias{statcast_get_coltypes}
\title{Obtain vector of expected Statcast column types}
\usage{
statcast_get_coltypes()
}
\value{
A character vector of the expected Statcast column types
}
\description{
Obtain vector of expected Statcast column types
}
