% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast-helpers.R
\name{statcast_make_url}
\alias{statcast_make_url}
\title{Make Statcast URL}
\usage{
statcast_make_url(date, batter, pitcher)
}
\arguments{
\item{date}{Game date}

\item{batter}{Statcast player ID for batter of interest. Defaults to \code{NULL}.}

\item{pitcher}{Statcast player ID for pitcher of interest. Defaults to \code{NULL}.}
}
\value{
A length one character vector contain a Statcast URL
}
\description{
Make Statcast URL
}
