% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_min_process}
\alias{statcast_min_process}
\title{Minimally process Statcast data}
\usage{
statcast_min_process(data)
}
\arguments{
\item{data}{A \code{data.frame}.}
}
\value{
A \code{data.frame}.
}
\description{
Minimally process Statcast data
}
