% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_names}
\alias{statcast_names}
\title{Add names to Statcast data}
\usage{
statcast_names(data)
}
\arguments{
\item{data}{A \code{data.frame}.}
}
\value{
An object with class \code{c("tbl_df", "tbl", "data.frame")}
}
\description{
Add names to Statcast data
}
