% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superficie_bbd.R
\name{superficie_bbd}
\alias{superficie_bbd}
\title{Gráfico de superfície para Box-Behnken}
\usage{
superficie_bbd(fit, x1, x2, n = 45)
}
\arguments{
\item{fit}{objeto da classe \code{bbd_fit}.}

\item{x1}{nome do primeiro fator.}

\item{x2}{nome do segundo fator.}

\item{n}{número de pontos da grade. Padrão \code{45}.}
}
\value{
Invisivelmente, uma lista com grade e matriz de predições.
}
\description{
Gera o gráfico de superfície de resposta para dois fatores do modelo ajustado,
mantendo os demais fatores fixados em zero.
}
