% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superficie_dbb.R
\name{superficie_dbb}
\alias{superficie_dbb}
\title{Grafico de superficie de resposta para Box Behnken}
\usage{
superficie_dbb(fit, x, y, z.fix = 0)
}
\arguments{
\item{fit}{objeto retornado por dbb_fit()}

\item{x}{nome do fator para eixo x}

\item{y}{nome do fator para eixo y}

\item{z.fix}{valor fixo do terceiro fator, default = 0}
}
\description{
Grafico de superficie de resposta para Box Behnken
}
