% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_efeitos_bbd.R
\name{tabela_efeitos_bbd}
\alias{tabela_efeitos_bbd}
\title{Tabela organizada dos efeitos do modelo BBD}
\usage{
tabela_efeitos_bbd(fit, alpha = 0.05)
}
\arguments{
\item{fit}{Objeto ajustado por \code{bbd_fit()}.}

\item{alpha}{Nivel de significancia. Padrao 0.05.}
}
\value{
data.frame com efeitos, erro padrao, p-valor e significancia.
}
\description{
Tabela organizada dos efeitos do modelo BBD
}
