% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_efeitos_dbb.R
\name{tabela_efeitos_dbb}
\alias{tabela_efeitos_dbb}
\title{Tabela organizada dos efeitos do modelo DBB}
\usage{
tabela_efeitos_dbb(fit)
}
\arguments{
\item{fit}{objeto retornado por dbb_fit()}
}
\value{
data.frame ordenado por magnitude do efeito
}
\description{
Tabela organizada dos efeitos do modelo DBB
}
