% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{two_cols_identical}
\alias{two_cols_identical}
\title{Check if two columns of a data frame are identical}
\usage{
two_cols_identical(df, col_1 = 1, col_2 = 2)
}
\arguments{
\item{df}{A data frame}

\item{col_1}{The index or name of a column}

\item{col_2}{The index or name of a column}
}
\value{
A logical vector
}
\description{
Check if two columns of a data frame are identical
}
