% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-reference.R
\name{bref_war_daily_pitcher}
\alias{bref_war_daily_pitcher}
\title{Download daily updated WAR data for pitchers from Baseball-Reference}
\usage{
bref_war_daily_pitcher(tibble = TRUE)
}
\arguments{
\item{tibble}{Controls class of object returned. Defaults to \code{TRUE} which
returns a \code{tibble}. When \code{FALSE}, returns a \code{data.frame}.}
}
\value{
A \code{data.frame} or \code{tibble} containing pitcher WAR data.
}
\description{
Download daily updated WAR data for pitchers from Baseball-Reference
}
