% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar_excel_bbd.R
\name{exportar_excel_bbd}
\alias{exportar_excel_bbd}
\title{Exporta resultados do BBD para Excel}
\usage{
exportar_excel_bbd(
  fit,
  arquivo = NULL,
  usar_desktop = TRUE,
  alpha = 0.05,
  fatores = NULL,
  objetivo = "max"
)
}
\arguments{
\item{fit}{objeto da classe bbd_fit}

\item{arquivo}{nome do arquivo Excel}

\item{usar_desktop}{se TRUE, salva na Área de Trabalho, em uma pasta
chamada BBD_Resultados}

\item{alpha}{nível de significância para destacar efeitos}

\item{fatores}{vetor opcional com os fatores a considerar nos gráficos;
se NULL, usa fit$fatores}

\item{objetivo}{objetivo da otimização: "max" para maximizar ou "min"
para minimizar. Padrão \code{"max"}.}
}
\value{
invisivelmente, o caminho do arquivo gerado
}
\description{
Exporta os resultados numéricos do planejamento Box-Behnken para Excel,
incluindo dados, métricas, ANOVA, coeficientes, efeitos, ponto ótimo
e ponto estacionário.
}
