% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar_excel_bdd.R
\name{exportar_excel_bbd}
\alias{exportar_excel_bbd}
\title{Exporta resultados do BBD para Excel com abas e gráficos}
\usage{
exportar_excel_bbd(
  fit,
  arquivo = "relatorio_bbd.xlsx",
  usar_desktop = TRUE,
  alpha = 0.05,
  fatores = NULL
)
}
\arguments{
\item{fit}{objeto da classe bbd_fit}

\item{arquivo}{nome do arquivo Excel}

\item{usar_desktop}{se TRUE, salva na Área de Trabalho, em uma pasta
chamada BBD_Resultados}

\item{alpha}{nível de significância para destacar efeitos}

\item{fatores}{vetor opcional com os fatores a considerar nos gráficos;
se NULL, usa fit$fatores}
}
\value{
invisivelmente, o caminho do arquivo gerado
}
\description{
Exporta resultados do BBD para Excel com abas e gráficos
}
