% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbam.R
\name{mlb_roster}
\alias{mlb_roster}
\title{Get MLB active rosters}
\usage{
mlb_roster(team)
}
\arguments{
\item{team}{A numeric MLB statsapi team ID.}
}
\value{
A data frame contain a team's active roster
}
\description{
Get MLB active rosters
}
