% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast-helpers.R
\name{statcast_min_process}
\alias{statcast_min_process}
\title{Minimally process Statcast data}
\usage{
statcast_min_process(data)
}
\arguments{
\item{data}{A \code{data.frame} with columns that match the return of \code{statcast_get_col_types()}.}
}
\value{
A \code{data.frame} containing minimally processed Statcast data
}
\description{
Minimally process Statcast data
}
