% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_names}
\alias{statcast_names}
\title{Add names to Statcast data}
\usage{
statcast_names(data, tibble = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{tibble}{Controls class of object returned. Defaults to \code{TRUE} which
returns a \code{tibble}. When \code{FALSE}, returns a \code{data.frame}.}
}
\value{
A \code{data.frame} or \code{tibble}.
}
\description{
Add names to Statcast data
}
